+++ src/loader/data/json/types/CreepSpawnData.java
@@ -6,5 +6,5 @@ package loader.data.json.types;
 
 public class CreepSpawnData {
   public String creep;
-  public float  spawnTime;
+  public int spawnTime;
 }
